refactor(users): move use cases into user folder
